refact: move SerializableList back to its original package
