--- loadable/src/main/java/com/jeanbarrossilva/loadable/list/serializable/SerializableList.extensions.kt
@@ -1,20 +1,0 @@
-package com.jeanbarrossilva.loadable.list.serializable
-
-/**
- * Creates a [SerializableList] without elements.
- *
- * @see serializableListOf
- **/
-fun <T> emptySerializableList(): SerializableList<T> {
-    return serializableListOf()
-}
-
-/**
- * Creates a new [SerializableList] with the given [elements].
- *
- * @param elements Elements to be added to the [SerializableList].
- **/
-fun <T> serializableListOf(vararg elements: T): SerializableList<T> {
-    val elementsAsList = elements.toList()
-    return SerializableList(elementsAsList)
-}
