+++ loadable/src/test/java/com/jeanbarrossilva/loadable/list/SerializableListExtensions.kt
@@ -1,10 +1,9 @@
 package com.jeanbarrossilva.loadable.list
 
 import com.jeanbarrossilva.loadable.Loadable
-import com.jeanbarrossilva.loadable.list.serializable.SerializableList
-import com.jeanbarrossilva.loadable.list.serializable.emptySerializableList
-import com.jeanbarrossilva.loadable.list.serializable.serializableListOf
+import java.io.Serializable
 import kotlin.test.Test
+import kotlin.test.assertContentEquals
 import kotlin.test.assertIs
 
 internal class SerializableListExtensions {
@@ -27,5 +26,15 @@ internal class SerializableListExtensions {
         assertIs<ListLoadable.Populated<SerializableList<Int>>>(
             Loadable.Loaded(serializableListOf(1, 2, 3)).asListLoadable()
         )
+    }
+
+    @Test
+    fun `GIVEN a SerializableList created through emptySerializableList WHEN checking if it's empty THEN it is`() { // ktlint-disable max-line-length
+        assertContentEquals(emptyList(), emptySerializableList<Serializable>())
+    }
+
+    @Test
+    fun `GIVEN a SerializableList created through serializableListOf WHEN iterating through it THEN all the given elements are present`() { // ktlint-disable max-line-length
+        assertContentEquals(listOf("A", "B", "C"), serializableListOf("A", "B", "C"))
     }
 }
